Remove PostCSS
After a few months of development, most of the bugs in the site have
ended up being caused by PostCSS purging CSS that it shouldn't have.
This has been particularly annoying because many of these bugs were only
found in prod, since PostCSS doesn't run in development environments.

While PostCSS provides many options for dealing with this, the PostCSS
config was quickly becoming unmaintainable.

While we may be able to address this by radically refactoring the
codebase, for the time being, the benefit we get from PostCSS isn't
worth the added development pains.