Relax guidance on Wayback links
We probably don't need to be adding a Wayback link for every regular
link. If the site is ever taken down, users can just go to the Wayback
Machine themselves, or we can edit the artifact to add a Wayback link.

This commit updates the guidance to only instruct users to use Wayback
links when the site is no longer available on the web or has changed
substantially.