Change default sort order for artifacts
I didn't like sorting artifacts in descending order by year, because
it's the first thing users see when they come to the site, and I don't
want to only showcase more recent ace history.

I tried sorting alphabetically, but that's tricky because of
punctuation. What ends up happening is all the artifacts that start with
`"` come first.

I like the idea of having the sort order be pseudorandom, but I also
don't want it to shuffle on every page load, because that would be
jarring.

So, I decided on sorting by artifact ID. It's effectively random,
because artifact IDs are allocated randomly, but it's still stable
between page loads.