Re-add PurgeCSS
We previously removed PurgeCSS because necessary CSS was frequently
being stripped, and it was becoming a maintenance burden to fix.
However, after removing, the size of the CSS bundle was much too large
and was affecting performance.

For some reason, configuring PurgeCSS was much simpler this time. This
might have to do with refactoring that's occurred since we last used
PurgeCSS.