Make artifacts page an index page
Previously, the main artifacts list page showed a list of all artifacts,
sorted by their unique internal ID. The intention was to choose a
seemingly random sort order to avoid presenting any one artifact over
any other, since this is the default page users will land on.

However, the artifacts presented at the top of the list weren't
necessarily representative of the contents of the archive as a whole,
which could leave visitors with a bad first impression.

So, we've reworked the main artifacts page to just be a list of
categories that users can dig into to find what they're looking for.